added selector tests & moved selector_tests
