+++ tests/test_preprocessing/test_pandastypeselector.py
@@ -3,25 +3,6 @@ import pandas as pd
 import itertools as it
 import numpy as np
 from sklego.preprocessing import PandasTypeSelector
-from tests.conftest import id_func
-
-
-@pytest.mark.parametrize("transformer", [
-    PandasTypeSelector(include=['number']),
-], ids=id_func)
-def test_len_regression(transformer, random_xy_dataset_regr):
-    X, y = random_xy_dataset_regr
-    X = pd.DataFrame(X)
-    assert transformer.fit(X, y).transform(X).shape[0] == X.shape[0]
-
-
-@pytest.mark.parametrize("transformer", [
-    PandasTypeSelector(include=['number']),
-], ids=id_func)
-def test_len_classification(transformer, random_xy_dataset_clf):
-    X, y = random_xy_dataset_clf
-    X = pd.DataFrame(X)
-    assert transformer.fit(X, y).transform(X).shape[0] == X.shape[0]
 
 
 @pytest.mark.parametrize('include,exclude', [_ for _ in it.combinations(['number', 'datetime', 'timedelta',
